refactor(viewer): modify EditPointer file location to Annotation
