+++ libs/insight-viewer/src/Annotation/components/EditPointer/index.tsx
@@ -1,8 +1,7 @@
-/* eslint-disable no-nested-ternary */
 import React from 'react'
-import { EDIT_CIRCLE_RADIUS } from '../../const'
-import { editPointerStyle } from '../../Annotation/viewer.styles'
-import type { CursorStatus, EditMode, Point } from '../../types'
+import { EDIT_CIRCLE_RADIUS } from '../../../const'
+import { editPointerStyle } from '../../viewer.styles'
+import type { CursorStatus, EditMode, Point } from '../../../types'
 
 interface EditPointerProps {
   editPoint: Point
